[Refactor] Rename to `OfficialCongestionResponse.java`
